Updated mixins configuration, improved dynamic resources handlers and implemented mixin extras library
This commit updates the configuration of mixins and removes the refmap attribute from the mixins JSON files. This focuses on improving the dynamic resources handlers' code for both the client and the server. We also implemented the mixin extras library, which is reflected in the changes made to the build.gradle files across the project. Minor changes include the modification of method signature in 'WoodTypeMixin.java' and refactoring of some class imports and usage in a few other files.